mp: plot vs steps instead of time

+++ ttt/src/agents/alphazero_mp.py
@@ -294,7 +294,7 @@ class LearnerMetrics(TypedDict):
     type: str
     policy_loss: float
     value_loss: float
-    steps_trained: float
+    steps_trained: int
     steps_per_sec: float
     batches_per_sec: float
